Remap flush promises.
Resolves a race condition where a flush is sent due to a collapse and
then a subsequent government arrives with an empty map before the flush
arrives. We were not aware that we'd flushed and attempted to remap our
outstanding messages with the empty map.

Remapping the flush gives us something to do instead of barreling on
ahead with the outstanding message promise remap.

No tests yet. I'm probably going to redo them all.