Replace Monotonic with Paxos.
We where using the promise comparison functions in Monotonic. I've
wanted to move away from Monotonic now that JavaScript has `BigInt`.
I've only moved so far away. Keeping the hex value pair, but
implementing the comaparison as `BigInt`. Retaining only bits of
Monotonic that I was using in for Paxos.

Closes #80.